Safari-611.1.10/LayoutTests/accessibility/mac/attributed-string/attributed-string-for-range-with-options-expected.txt

word mispelled word
This test ensures that attributed string for text marker range works and only includes misspelled attribute when requested.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text is expectedOne
PASS text is expectedTwo
PASS successfullyParsed is true

TEST COMPLETE
